welcome: Fix greeting message crashing when clean service was on
A bug was causing Haruka not to be able to construct the welcome message correctly and falling back to backup welcome message, because clean service was deleting the service messages, which Haruka needed to reply to. Not being able to find this message, unexpected behavior was happening. This commit fixes that bug by making full use of the variable clean service provides, and we can tell Telegram not to reply to anything if clean service deleted the service messages.

Signed-off-by: Gegham Zakaryan <zakaryan.2004@outlook.com>